3.1-drf-intro/smart_home/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
